// src/email-marketing/enums/trigger-type.enum.ts
export enum TriggerType {
    USER_SIGNUP = 'user_signup',
    COURSE_ENROLLED = 'course_enrolled',
    COURSE_COMPLETED = 'course_completed',
    PURCHASE_MADE = 'purchase_made',
    USER_INACTIVE = 'user_inactive',
    SUBSCRIPTION_CREATED = 'subscription_created',
    SUBSCRIPTION_CANCELLED = 'subscription_cancelled',
    BIRTHDAY = 'birthday',
    CUSTOM_EVENT = 'custom_event',
    DATE_BASED = 'date_based',
    SEGMENT_ENTERED = 'segment_entered',
    SEGMENT_LEFT = 'segment_left',
}