// src/email-marketing/enums/action-type.enum.ts
export enum ActionType {
    SEND_EMAIL = 'send_email',
    WAIT = 'wait',
    ADD_TAG = 'add_tag',
    REMOVE_TAG = 'remove_tag',
    ADD_TO_SEGMENT = 'add_to_segment',
    REMOVE_FROM_SEGMENT = 'remove_from_segment',
    UPDATE_PROPERTY = 'update_property',
    WEBHOOK = 'webhook',
    SEND_SMS = 'send_sms',
    SEND_PUSH_NOTIFICATION = 'send_push_notification',
}